+++ libpay-worldpay-enterprise-testkit/src/main/scala/com/wix/pay/worldpay/enterprise/testkit/WorldpayDriver.scala
@@ -1,5 +1,4 @@
-package com.wix.pay.worldpay.testkit
-
+package com.wix.pay.worldpay.enterprise.testkit
 
 import java.math.{BigDecimal => JBigDecimal}
