+++ libpay-worldpay-enterprise-testkit/src/main/scala/com/wix/pay/worldpay/enterprise/testkit/XMLNoDTDValidation.scala
@@ -1,4 +1,4 @@
-package com.wix.pay.worldpay.testkit
+package com.wix.pay.worldpay.enterprise.testkit
 
 import javax.xml.parsers.SAXParser
