+++ libpay-worldpay-enterprise/src/main/scala/com/wix/pay/worldpay/enterprise/WorldpayAuthorization.scala
@@ -1,3 +1,3 @@
-package com.wix.pay.worldpay
+package com.wix.pay.worldpay.enterprise
 
 case class WorldpayAuthorization(orderCode: String, currency: String)
