+++ libpay-worldpay-enterprise/src/main/scala/com/wix/pay/worldpay/enterprise/parsers/JsonWorldpayAuthorizationParser.scala
@@ -1,6 +1,6 @@
-package com.wix.pay.worldpay.parsers
+package com.wix.pay.worldpay.enterprise.parsers
 
-import com.wix.pay.worldpay.WorldpayAuthorization
+import com.wix.pay.worldpay.enterprise.WorldpayAuthorization
 import org.json4s.DefaultFormats
 import org.json4s.native.Serialization
